extract ContractStateVaultJsonFactoryProvider & ContractStateVaultJsonFactoryRegistry interface
