+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/json/impl/ContractStateVaultJsonFactoryProvider.kt
@@ -1,6 +1,6 @@
 package net.corda.ledger.persistence.json.impl
 
-import net.corda.ledger.persistence.json.ContractStateVaultJsonFactoryRegistry
+import net.corda.ledger.libs.json.ContractStateVaultJsonFactoryRegistry
 import net.corda.sandbox.type.SandboxConstants
 import net.corda.sandbox.type.UsedByPersistence
 import net.corda.sandboxgroupcontext.CustomMetadataConsumer
