+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/json/impl/ContractStateVaultJsonFactoryRegistryImpl.kt
@@ -1,6 +1,6 @@
 package net.corda.ledger.persistence.json.impl
 
-import net.corda.ledger.persistence.json.ContractStateVaultJsonFactoryRegistry
+import net.corda.ledger.libs.json.ContractStateVaultJsonFactoryRegistry
 import net.corda.sandbox.type.UsedByPersistence
 import net.corda.v5.ledger.utxo.ContractState
 import net.corda.v5.ledger.utxo.query.json.ContractStateVaultJsonFactory
