+++ libs/persistence/src/main/java/net/corda/ledger/libs/json/package-info.java
@@ -1,0 +1,6 @@
+@Export
+@QuasarIgnoreAllPackages
+package net.corda.ledger.libs.json;
+
+import co.paralleluniverse.quasar.annotations.QuasarIgnoreAllPackages;
+import org.osgi.annotation.bundle.Export;
